not working (have to change the MainThread un() method)

+++ src/element/Ghost.java
@@ -11,91 +11,91 @@ import service.NodeService;
 
 public class Ghost extends AbstractPersonnage {
 
-    private static List<Ghost> ghosts;
+	private static List<Ghost> ghosts;
 
-    public static List<Ghost> getGhosts() {
-        return ghosts;
-    }
+	public static List<Ghost> getGhosts() {
+		return ghosts;
+	}
 
-    public static void setGhosts(List<Ghost> ghosts) {
-        Ghost.ghosts = ghosts;
-    }
+	public static void setGhosts(List<Ghost> ghosts) {
+		Ghost.ghosts = ghosts;
+	}
 
-    private GhostService ghs;
+	private GhostService ghs;
 
-    private NodeService ns;
+	private NodeService ns;
 
-    private List<AbstractNode> chemin;
+	private List<AbstractNode> chemin;
 
-    private AbstractNode currentNode;
+	private AbstractNode currentNode;
 
-    public Ghost(int x, int y, int width, int height) {
-        super(x, y, width, height);
-        ns = NodeService.getInstance();
-        ghs = GhostService.getInstance();
-        if (ghosts == null)
-            ghosts = new ArrayList<>();
-        ghosts.add(this);
-        chemin = new ArrayList<>();
-        currentNode = ns.getNodeByPos(this.actualPos);
-    }
+	public Ghost(int x, int y, int width, int height) {
+		super(x, y, width, height);
+		ns = NodeService.getInstance();
+		ghs = GhostService.getInstance();
+		if (ghosts == null)
+			ghosts = new ArrayList<>();
+		ghosts.add(this);
+		chemin = new ArrayList<>();
+		currentNode = ns.getNodeByPos(this.actualPos);
+	}
 
-    private void findDirection() {
-        if (!this.chemin.isEmpty()) {
-            AbstractNode target = this.chemin.get(0);
-            if (!target.getPos().equals(this.actualPos)) {
-                if (target.getPos().height == this.actualPos.height) {
-                    if (target.getPos().width > this.actualPos.width) {
-                        resetDirection(false);
-                        right = true;
-                    } else if (target.getPos().width < this.actualPos.width) {
-                        resetDirection(false);
-                        left = true;
-                    }
-                } else if (target.getPos().width == this.actualPos.width) {
-                    if (target.getPos().height > this.actualPos.height) {
-                        resetDirection(false);
-                        down = true;
-                    } else if (target.getPos().height < this.actualPos.height) {
-                        resetDirection(false);
-                        up = true;
-                    }
-                }
-            } else {
-                this.currentNode = target;
-                resetDirection(false);
-                this.chemin.remove(0);
-            }
-        }
-    }
+	private void findDirection() {
+		if (!this.chemin.isEmpty()) {
+			AbstractNode target = this.chemin.get(0);
+			if (!target.getPos().equals(this.actualPos)) {
+				if (target.getPos().height == this.actualPos.height) {
+					if (target.getPos().width > this.actualPos.width) {
+						resetDirection(false);
+						right = true;
+					} else if (target.getPos().width < this.actualPos.width) {
+						resetDirection(false);
+						left = true;
+					}
+				} else if (target.getPos().width == this.actualPos.width) {
+					if (target.getPos().height > this.actualPos.height) {
+						resetDirection(false);
+						down = true;
+					} else if (target.getPos().height < this.actualPos.height) {
+						resetDirection(false);
+						up = true;
+					}
+				}
+			} else {
+				this.currentNode = target;
+				resetDirection(false);
+				this.chemin.remove(0);
+			}
+		}
+	}
 
-    public List<AbstractNode> getChemin() {
-        return this.chemin;
-    }
+	public List<AbstractNode> getChemin() {
+		return this.chemin;
+	}
 
-    public AbstractNode getCurrentNode() {
-        return this.currentNode;
-    }
+	public AbstractNode getCurrentNode() {
+		return this.currentNode;
+	}
 
-    @Override
-    public void move() {
-        this.actualPos = gs.getPositionFromPixel(new Dimension(this.anchor.width, this.anchor.height));
-        findDirection();
-        direction();
-    }
+	@Override
+	public void move() {
+		this.actualPos = gs.getPositionFromPixel(new Dimension(this.anchor.width, this.anchor.height));
+		findDirection();
+		direction();
+	}
 
-    @Override
-    public void paint(Graphics g) {
-        g.setColor(Color.PINK);
-        g.fillOval(this.x, this.y, width, height);
-    }
+	@Override
+	public void paint(Graphics g) {
+		g.setColor(Color.PINK);
+		g.fillOval((int) this.x, (int) this.y, width, height);
+	}
 
-    public void setChemin(List<AbstractNode> chemin) {
-        this.chemin = chemin;
-    }
+	public void setChemin(List<AbstractNode> chemin) {
+		this.chemin = chemin;
+	}
 
-    public void setCurrentNode(AbstractNode currentNode) {
-        this.currentNode = currentNode;
-    }
+	public void setCurrentNode(AbstractNode currentNode) {
+		this.currentNode = currentNode;
+	}
 
 }
